+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## 0.23.0 - 2025-10-20
 ### Changed
 - Upgrae Rust edition to 2024.
 - [**BREAKING**] When `default` is an expression that uses previous fields,